pyproject.toml
[tool.poetry]
name = "king-gizzard-setlist-predictor"
version = "0.1.0"
description = "King Gizzard & The Lizard Wizard Setlist Predictor using GNN and Deep Learning"
authors = ["Junichi Koganemaru"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
# Core data processing
pandas = "^2.2.0"
numpy = "^1.26.0"
pyarrow = "^15.0.0"
python-dotenv = "^1.0.1"
pydantic = "^2.6.0"
pydantic-settings = "^2.1.0"
# Machine learning
scikit-learn = "^1.4.0"
xgboost = "^2.0.3"
torch = "^2.0.0"
torch-geometric = "^2.4.0"
torch-scatter = "^2.1.0"
torch-sparse = "^0.6.0"
# Hyperparameter tuning
optuna = "^3.5.0"
# Utilities
requests = "^2.31.0"
diskcache = "^5.6.3"
joblib = "^1.3.2"
tqdm = "^4.66.1"
click = "^8.1.7"
python-dateutil = "^2.8.2"
pyyaml = "^6.0.1"
# Visualization
matplotlib = "^3.8.2"
seaborn = "^0.13.2"
scipy = "^1.12.0"
networkx = "^3.2.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
black = "^24.1.1"
ruff = "^0.2.0"
mypy = "^1.8.0"
pre-commit = "^3.6.0"
autopep8 = "^2.0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v"
